refactor logic code
