+++ CSXS/manifest.xml
@@ -39,7 +39,7 @@
 				</Lifecycle>
 				<UI>
 					<Type>Panel</Type>
-					<Menu>Plainly Videos - Dev mode</Menu>
+					<Menu>Plainly Videos</Menu>
 					<Geometry>
 						<Size>	
 							<Height>540</Height>
